refactor: refactoring user-feature, item-feature, добавлен предварительный каркас для itemRequest and booking
